Updates incliude: (SearchCard components) - create components/SearchCard folder and move its related components(SearchCard, SearchCardHeader, SearchCardMeta, SearchCardBody, SearchCardFooter) - rename SearchCard.js file to SearchCard/index.js - remove Wrappers and styled-component imports from SearchCard(index.js), SearchHeader, and SearchBody components - remove the sub component SearchCardTitle and add its JSX directly in SearchCardHeader - (in progress) add a sub-component DatasetCTA in SearchCardHeader and move all conditionally rendered JSX based on a 'status'
