+++ src/components/SearchCard/SearchCardFooter.js
